pyproject.toml
[tool.poetry]
packages = [{ include = "src" }]
name = "bot-secretar"
version = "3.0.0"
description = "telegram bot"
authors = ["A-V-tor <avtorca4@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.21.0"
pillow = "10.0.0"
flask-login = "^0.6.2"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
prettytable = "^3.5.0"
alembic = "^1.11.1"
gunicorn = "^21.2.0"
redis = "^5.0.1"
dash = "^2.14.2"
aiogram = "^3.13.1"
pandas = "^2.2.3"
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.36"
wtforms = "3.0.1"
flask-babel = "^4.0.0"
flask = "3.0.0"
flask-ckeditor = "^1.0.0"
bleach = "^6.2.0"
flask-admin = "2.0.0a1"
python-nginx = "^1.5.7"
apscheduler = "^3.11.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.19.4"
pre-commit = "^3.6.0"
ruff = "^0.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bot = "main:main"

[tool.ruff]
line-length = 120
fix = true

[tool.ruff.format]
docstring-code-format = true
quote-style = "single"

[tool.ruff.lint]
ignore = []
select = ["I"]

[tool.ruff.lint.isort]
force-wrap-aliases = true
combine-as-imports = true


[tool.pyright]
venvPath = "."
venv = ".venv"

include = ["src"]
exclude = ["**/__pycache__"]

pythonVersion = "3.11"
